+++ app/controllers/randomizations_controller.rb
@@ -1,14 +1,24 @@
 class RandomizationsController < ApplicationController
   before_action :authenticate_user!
-  before_action :set_editable_project
+
+  before_action :set_viewable_project,                  only: [:index, :show]
+  before_action :set_editable_project_or_editable_site, only: [:choose_scheme, :undo]
   before_action :redirect_without_project
-  before_action :set_randomization,                only: [:show, :undo, :destroy]
-  before_action :redirect_without_randomization,   only: [:show, :undo, :destroy]
+
+  before_action :set_viewable_randomization,            only: [:show]
+  before_action :set_editable_randomization,            only: [:undo] # , :destroy
+  before_action :redirect_without_randomization,        only: [:show, :undo] # , :destroy
+
+  def choose_scheme
+    if @project.randomization_schemes.published.count == 1
+      redirect_to randomize_subject_to_list_project_randomization_scheme_path(@project, @project.randomization_schemes.published.first)
+    end
+  end
 
   # GET /randomizations
   # GET /randomizations.json
   def index
-    @randomizations = @project.randomizations.includes(:subject).order("randomized_at DESC NULLS LAST").page(params[:page]).per(40)
+    @randomizations = current_user.all_viewable_randomizations.where(project_id: @project.id).includes(:subject).order("randomized_at DESC NULLS LAST").page(params[:page]).per(40)
   end
 
   # GET /randomizations/1
@@ -37,8 +47,12 @@ class RandomizationsController < ApplicationController
   # end
 
   private
-    def set_randomization
-      @randomization = @project.randomizations.find_by_id(params[:id])
+    def set_viewable_randomization
+      @randomization = current_user.all_viewable_randomizations.find_by_id(params[:id])
+    end
+
+    def set_editable_randomization
+      @randomization = current_user.all_randomizations.find_by_id(params[:id])
     end
 
     def redirect_without_randomization