+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - Two types of randomization algorithms are available to use
   - Randomization criteria are now stored when subject is randomized
   - When site is used as a stratification factor, subjects can only be stratified to their own site
+  - Site Editors can create and undo randomizations for subjects on their site
+  - Site Viewers can view all existing randomization for subjects on their site
   - **Permuted-block Algorithm**
     - Treatment Arms can now be specified along with a weight allocation
     - Different block sizes and allocations can now be specified